Changes to the asset management (prevent exception), fixes on the hierarchy (now the first iteration will show root elements only)

+++ GoldEngine-Internal/Scene.h
@@ -85,7 +85,8 @@ namespace Engine::Management
 
 			sceneDatapack->ReadFromFile(sceneRequirements, password);
 
-			Directory::Delete("Data/tmp/", true);
+			if(Directory::Exists("Data/tmp/"))
+				Directory::Delete("Data/tmp/", true);
 
 			sceneFinishedLoading = true;
